Content container display fixed
The size of container depends on amount of elements inside, thus it is extended when elements are overflowing. But if elements are not enough, content container doesn't reach bottom of the page. The solutions is to create an image and attach it to `body` background (as in original theme it is done) with the same size and color as container ones. Setting `repeat-y` it will fit the entire page from the top to the bottom.